+++ KafkaToFileConnector/src/main/java/com/citi/stg/KafkaToFileConnector/stringToFileThread/stringToFile.java
@@ -3,10 +3,16 @@ package com.citi.stg.KafkaToFileConnector.stringToFileThread;
 
 import com.citi.stg.KafkaToFileConnector.ConverterService.Converter;
 
+/*
+ * Thread class  calling the main service of the microservice.
+ *
+ * Multithreading is used to achieve parallel processing capability.
+ */
 public class stringToFile extends Thread {
     private String xmlString;
-    private int i;
+    private long i;
 
+    // constructor
     public stringToFile(String xmlString, int i) {
         this.xmlString = xmlString;
         this.i = i;
